+++ staging_service/autodetect/Mappings.py
@@ -9,7 +9,7 @@ ZIP = "CompressedFileFormatArchive"
 
 # BIOINFORMATICS FORMATS
 FASTA = "FASTA"
-FASTQ = "FASTQ Reads"
+FASTQ = "FASTQ"
 GFF = "GFF"
 GTF = "GTF"
 SRA = "SRA"
@@ -19,9 +19,9 @@ GENBANK = "GENBANK"
 VCF = "VCF"
 FBA = "FBAModel"
 SBML = "SBML"
+MSA = "MSA"
 
 # KBASE SPECIFIC FORMATS
-MSA = "MultipleSequenceAlignment"
 PHENOTYPE = "PHENOTYPE"  # "KBasePhenotypes.PhenotypeSet"
 ESCHER = "ESCHER"
 ANNOTATIONS = "ANNOTATIONS"
@@ -55,13 +55,13 @@ escher_map_id = "escher_map"
 def _flatten(some_list):
     return list(itertools.chain.from_iterable(some_list))
 
-_GZIP_EXT = ["", ".gz", ".gzip"]  # empty string to keep the uncompressed extension
+_COMPRESSION_EXT = ["", ".gz", ".gzip"]  # empty string to keep the uncompressed extension
 
 # longer term there's probably a better way to do this but this is quick
 def _add_gzip(extension_list):
-    return _flatten([[ext + gz for gz in _GZIP_EXT] for ext in extension_list])
+    return _flatten([[ext + comp for comp in _COMPRESSION_EXT] for ext in extension_list])
 
-type_to_extension_mapping = {
+file_format_to_extension_mapping = {
     FASTA: _add_gzip(["fna", "fa", "faa", "fsa", "fasta"]),
     FASTQ: _add_gzip(["fq", "fastq"]),
     GFF: _add_gzip(["gff", "gff2", "gff3"]),
@@ -91,10 +91,12 @@ type_to_extension_mapping = {
     EXCEL: ["xls", "xlsx"],
     ZIP: ["zip", "tar", "tgz", "tar.gz", "7z", "gz", "gzip", "rar"],
     SBML: ["smbl"],
-    # Custom File Types?
-    MEDIA: ["tsv", "xls", "xlsx"],
-    PHENOTYPE: ["tsv"],
-    ESCHER: ["json"],
-    ANNOTATIONS: ["tsv"],
-    FBA: ["tsv", "xls", "xlsx", "smbl"],
 }
+
+extension_to_file_format_mapping = {}
+for type_, extensions in file_format_to_extension_mapping.items():
+    for ext in extensions:
+        if ext in extension_to_file_format_mapping:
+            type2 = extension_to_file_format_mapping[ext]
+            raise ValueError(f"Duplicate entry for extension {ext} in {type_} and {type2}")
+        extension_to_file_format_mapping[ext] = type_